+++ affiliate/types/doc.go
@@ -1,0 +1,5 @@
+/*
+FILE: affiliate/types/doc.go
+*/
+
+package types